Add check to resize each file only once
Because we want to check if an image exceeds the target dimensions,
the way dimensions are supplied has changed. It is now only possible
to submit a target width and height instead of a dimension. This
means resizing by 50% for example will not work anymore. This action
needs specific pixels.